requirements.txt
aniso8601==9.0.1
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==4.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.1
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.8
l18n==2021.3
laces==0.1.1
Markdown==3.7
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pillow_heif==0.18.0
promise==2.3
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
text-unidecode==1.3
urllib3==2.2.2
wagtail==6.2.1
wagtail-markdown==0.12.1
webencodings==0.5.1
Willow==1.8.0
